refactor: 내용 폼 분리
